requirements.txt

streamlit
Pillow
qrcode[pil]
reportlab
